// model/student.ts
export class Student {
  loginName: string;
  stdId: string;
  stdCode: string;
  titleTh: string;
  titleEn: string;
  firstNameTh: string;
  middleNameTh: string;
  lastNameTh: string;
  firstNameEn: string;
  middleNameEn: string;
  lastNameEn: string;
  copenTd: string;
  copenNameTh: string;
  copenNameEn: string;
  campusCode: string;
  campusNameTh: string;
  campusNameEn: string;
  facultyCode: string;
  facultyNameTh: string;
  facultyNameEn: string;
  departmentCode: string;
  departmentNameTh: string;
  departmentNameEn: string;
  majorCode: string;
  magorNameTh: string;
  magorNameEn: string;
  nationCode: string;
  nationalityNameTh: string;
  nationalityNameEn: string;
  studentStatusCode: string;
  studentStatusNameTh: string;
  studentStatusNameEn: string;
  studentTypeCode: string;
  studentTypeNameTh: string;
  studentTypeNameEn: string;
  edulevelCode: string;
  edulevelNameTh: string;
  edulevelNameEn: string;
  studentYear: string;
  advisorId: string;
  advisorNameTh: string;
  advisorNameEn: string;
  positionTh: string;
  email: string;
  mobileNo: string;

  constructor(secionJson: any) {
    this.loginName = secionJson["loginName"];
    this.stdId = secionJson["stdId"];

    this.stdCode = secionJson["stdCode"];
    this.titleTh = secionJson["titleTh"];
    this.titleEn = secionJson["titleEn"];
    this.firstNameTh = secionJson["firstNameTh"];
    this.middleNameTh = secionJson["middleNameTh"];
    this.lastNameTh = secionJson["lastNameTh"];
    this.firstNameEn = secionJson["firstNameEn"];
    this.middleNameEn = secionJson["middleNameEn"];
    this.lastNameEn = secionJson["lastNameEn"];
    this.copenTd = secionJson["copenTd"];
    this.copenNameTh = secionJson["copenNameTh"];
    this.copenNameEn = secionJson["copenNameEn"];
    this.campusCode = secionJson["campusCode"];
    this.campusNameTh = secionJson["campusNameTh"];
    this.campusNameEn = secionJson["campusNameEn"];
    this.facultyCode = secionJson["facultyCode"];
    this.facultyNameTh = secionJson["facultyNameTh"];
    this.facultyNameEn = secionJson["facultyNameEn"];
    this.departmentCode = secionJson["departmentCode"];
    this.departmentNameTh = secionJson["departmentNameTh"];
    this.departmentNameEn = secionJson["departmentNameEn"];
    this.majorCode = secionJson["majorCode"];
    this.magorNameTh = secionJson["magorNameTh"];
    this.magorNameEn = secionJson["magorNameEn"];
    this.nationCode = secionJson["nationCode"];
    this.nationalityNameTh = secionJson["nationalityNameTh"];
    this.nationalityNameEn = secionJson["nationalityNameEn"];
    this.studentStatusCode = secionJson["studentStatusCode"];
    this.studentStatusNameTh = secionJson["studentStatusNameTh"];
    this.studentStatusNameEn = secionJson["studentStatusNameEn"];
    this.studentTypeCode = secionJson["studentTypeCode"];
    this.studentTypeNameTh = secionJson["studentTypeNameTh"];
    this.studentTypeNameEn = secionJson["studentTypeNameEn"];
    this.edulevelCode = secionJson["edulevelCode"];
    this.edulevelNameTh = secionJson["edulevelNameTh"];
    this.edulevelNameEn = secionJson["edulevelNameEn"];
    this.studentYear = secionJson["studentYear"];
    this.advisorId = secionJson["advisorId"];
    this.advisorNameTh = secionJson["advisorNameTh"];
    this.advisorNameEn = secionJson["advisorNameEn"];
    this.positionTh = secionJson["positionTh"];
    this.email = secionJson["email"];
    this.mobileNo = secionJson["mobileNo"];
  }
}